+++ apps/audience/src/pages/home/home.tsx
@@ -1,13 +1,8 @@
-import { useEffect, useRef, useState } from 'react';
 import { useQuery } from '@tanstack/react-query';
 import { useNavigate } from 'react-router';
 
 import { HomeBanner, InstallGuideSheet } from '@amp/compositions';
-import {
-  dismissInstallGuideForToday,
-  shouldShowInstallGuide,
-} from '@amp/shared/utils';
-import { getMobileOs } from '@amp/shared/utils';
+import { usePwaInstallGuide } from '@amp/shared/hooks';
 
 import FestivalSection from '@widgets/home/components/festival-section/festival-section';
 
@@ -19,35 +14,13 @@ import useHomeFestivals from './model/use-home-festivals';
 
 import { page } from './home.css';
 
-interface BeforeInstallPromptEvent extends Event {
-  prompt: () => Promise<void>;
-  userChoice: Promise<{
-    outcome: 'accepted' | 'dismissed';
-    platform: string;
-  }>;
-}
-
 const HomePage = () => {
   const navigate = useNavigate();
-  const deferredPromptRef = useRef<BeforeInstallPromptEvent | null>(null);
 
-  const [isOpen, setIsOpen] = useState(() => shouldShowInstallGuide());
-
-  useEffect(() => {
-    const handleBeforeInstallPrompt = (event: Event) => {
-      event.preventDefault();
-      deferredPromptRef.current = event as BeforeInstallPromptEvent;
-    };
-
-    window.addEventListener('beforeinstallprompt', handleBeforeInstallPrompt);
-
-    return () => {
-      window.removeEventListener(
-        'beforeinstallprompt',
-        handleBeforeInstallPrompt,
-      );
-    };
-  }, []);
+  const { isOpen, handleOpenApp, handleBrowseToday, handleClose } =
+    usePwaInstallGuide({
+      onMoveToGuide: () => navigate(ROUTE_PATH.PWA_GUIDE),
+    });
 
   const { data } = useQuery({
     ...USER_QUERY_OPTIONS.NICKNAME(),
@@ -64,34 +37,6 @@ const HomePage = () => {
 
   const handleCardClick = (festivalId: number) => {
     navigate(NAV_PATH.noticeList(festivalId));
-  };
-
-  const handleOpenApp = async () => {
-    const os = getMobileOs();
-
-    if (os === 'ios') {
-      navigate(ROUTE_PATH.PWA_GUIDE);
-      return;
-    }
-
-    if (os === 'android' && deferredPromptRef.current) {
-      await deferredPromptRef.current.prompt();
-      await deferredPromptRef.current.userChoice;
-      deferredPromptRef.current = null;
-      setIsOpen(false);
-      return;
-    }
-
-    navigate(ROUTE_PATH.PWA_GUIDE);
-  };
-
-  const handleBrowseToday = () => {
-    dismissInstallGuideForToday();
-    setIsOpen(false);
-  };
-
-  const handleClose = () => {
-    setIsOpen(false);
   };
 
   return (
