+++ packages/shared/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useDraggableCarousel } from './use-draggable-carousel';
 export { useNoticeList } from './use-notice-list';
+export { usePwaInstallGuide } from './use-pwa-install-guide';
